Extracted notification from DownloadService
